Added tests and debugged
