+++ pathaia/graphs/__init__.py
@@ -1,2 +1,2 @@
-from .graph import *
+from .object_api import *
 from .functional_api import *
