+++ tox.ini
@@ -15,6 +15,7 @@ envlist = py35, py36, py37, py38
 deps = 
     coverage
     pytest
+    ordered-set
 commands =
   coverage run --source pathaia -m pytest tests --verbose
   coverage xml
